# Batch verify aggregated attestation signatures