This PR batches verify all 3 signatures within aggregated attestation, the 3 signatures in aggregated attestation: selection signature, aggregator signature, and attestation signature. 

It combines the following 3 verifications into 1:
![image](https://user-images.githubusercontent.com/21316537/98453485-71f3e480-210e-11eb-9a8a-2b9620348cae.png)

![image](https://user-images.githubusercontent.com/21316537/98453512-c4cd9c00-210e-11eb-9f41-046d84363107.png)


With the following benchmark: https://gist.github.com/terencechain/412ef767ff37562a2735237c7ffc8b71
We gathered the following results:
```
Before:
BenchmarkValidateAggregateAndProof
BenchmarkValidateAggregateAndProof-12             295       3931547 ns/op
PASS

After:
BenchmarkValidateAggregateAndProof
BenchmarkValidateAggregateAndProof-12             414       2812799 ns/op
```

The result showed 40% improvements, roughly 1200000ns of saving on per attestation. There's currently 260 aggregated attestation per slot in medalla. That’s about 300ms we could save on per slot.